# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/api/countries/getByName":{"get":{"tags":["country-controller"],"operationId":"getCountriesName","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/api/countries/getByLenguage":{"get":{"tags":["country-controller"],"operationId":"getCountriesLenguage","parameters":[{"name":"len","in":"query","required":true,"schema":{"type":"string","enum":["English","Spanish","French","German","Portuguese","Chinese","Arabic","Russian","Hindi","Swahili"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/api/countries/getByContinent":{"get":{"tags":["country-controller"],"operationId":"getCountriesCon","parameters":[{"name":"con","in":"query","required":true,"schema":{"type":"string","enum":["Africa","Americas","Asia","Europe","Oceania","Antartic"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}},"/api/countries/getByCode":{"get":{"tags":["country-controller"],"operationId":"getCountriesCode","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}}}}}}}},"components":{"schemas":{"Country":{"type":"object","properties":{"name":{"type":"string"},"population":{"type":"integer","format":"int64"},"area":{"type":"number","format":"double"},"code":{"type":"string"},"region":{"type":"string"},"borders":{"type":"array","items":{"type":"string"}},"languages":{"type":"object","additionalProperties":{"type":"string"}}}}}}}